+++ server/src/main/java/com/inha/server/oauth/service/OAuthService.java
@@ -3,10 +3,9 @@ package com.inha.server.oauth.service;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.ObjectMapper;
-import com.inha.server.oauth.dto.JwtDto;
 import com.inha.server.oauth.dto.OAuthUserDto;
 import com.inha.server.oauth.model.KakaoTokens;
-import com.inha.server.security.jwt.TokenProvider;
+import com.inha.server.user.util.TokenProvider;
 import com.inha.server.user.model.User;
 import com.inha.server.user.repository.UserRepository;
 import java.time.LocalDateTime;
@@ -18,8 +17,6 @@ import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpMethod;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
-import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
-import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -45,7 +42,7 @@ public class OAuthService {
   @Value("${oauth.kakao.url.host}")
   private String redirectHost;
 
-  public JwtDto kakaoLogin(String authorize_code) throws JsonProcessingException {
+  public String kakaoLogin(String authorize_code) throws JsonProcessingException {
     // 인가코드를 통해서 access_token 발급
     String accessToken = getAccessToken(authorize_code);
     // access_token 으로 회원정보 받아오기
@@ -53,7 +50,7 @@ public class OAuthService {
     // 회원가입
     User user = signUpIfNotRegisteredUser(oAuthUserDto);
     // 로그인 후 jwt 리턴
-    return new JwtDto(login(user));
+    return login(user);
   }
 
   private String getAccessToken(String authorize_code) {
@@ -124,9 +121,6 @@ public class OAuthService {
   }
 
   private String login(User user) {
-    UsernamePasswordAuthenticationToken authenticationToken =
-        new UsernamePasswordAuthenticationToken(user.getId(), user.getPassword());
-    SecurityContextHolder.getContext().setAuthentication(authenticationToken);
-    return tokenProvider.createToken(authenticationToken);
+    return tokenProvider.createToken(user);
   }
 }
