--- server/src/main/java/com/inha/server/security/config/SecurityConfig.java
@@ -1,50 +1,0 @@
-package com.inha.server.security.config;
-
-import com.inha.server.security.exception.JwtAccessDeniedHandler;
-import com.inha.server.security.exception.JwtAuthenticationEntryPoint;
-import com.inha.server.security.jwt.JwtSecurityConfig;
-import com.inha.server.security.jwt.TokenProvider;
-import lombok.RequiredArgsConstructor;
-import org.springframework.context.annotation.Bean;
-import org.springframework.security.config.annotation.method.configuration.EnableGlobalMethodSecurity;
-import org.springframework.security.config.annotation.web.builders.HttpSecurity;
-import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
-import org.springframework.security.config.http.SessionCreationPolicy;
-import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
-import org.springframework.security.crypto.password.PasswordEncoder;
-import org.springframework.security.web.SecurityFilterChain;
-
-@EnableWebSecurity
-@EnableGlobalMethodSecurity(prePostEnabled = true)
-@RequiredArgsConstructor
-public class SecurityConfig {
-
-  private final TokenProvider tokenProvider;
-  private final JwtAccessDeniedHandler accessDeniedHandler;
-  private final JwtAuthenticationEntryPoint authenticationEntryPoint;
-
-  @Bean
-  public PasswordEncoder passwordEncoder() {
-    return new BCryptPasswordEncoder();
-  }
-
-  @Bean
-  public SecurityFilterChain filterChain(HttpSecurity httpSecurity) throws Exception {
-    httpSecurity
-        .csrf().disable()
-        .exceptionHandling()
-        .authenticationEntryPoint(authenticationEntryPoint)
-        .accessDeniedHandler(accessDeniedHandler)
-        .and()
-        .sessionManagement()
-        .sessionCreationPolicy(SessionCreationPolicy.STATELESS)
-        .and()
-        .authorizeRequests()
-        .antMatchers("/oauth/kakao").permitAll()
-        .anyRequest().authenticated()
-        .and()
-        .apply(new JwtSecurityConfig(tokenProvider));
-
-    return httpSecurity.build();
-  }
-}
