--- server/src/main/java/com/inha/server/security/exception/JwtAccessDeniedHandler.java
@@ -1,18 +1,0 @@
-package com.inha.server.security.exception;
-
-import java.io.IOException;
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-import org.springframework.security.access.AccessDeniedException;
-import org.springframework.security.web.access.AccessDeniedHandler;
-import org.springframework.stereotype.Component;
-
-@Component
-public class JwtAccessDeniedHandler implements AccessDeniedHandler {
-
-  @Override
-  public void handle(HttpServletRequest request, HttpServletResponse response,
-      AccessDeniedException accessDeniedException) throws IOException {
-    response.sendError(HttpServletResponse.SC_FORBIDDEN);
-  }
-}
