--- server/src/main/java/com/inha/server/security/jwt/JwtFilter.java
@@ -1,32 +1,0 @@
-package com.inha.server.security.jwt;
-
-import java.io.IOException;
-import javax.servlet.FilterChain;
-import javax.servlet.GenericFilter;
-import javax.servlet.ServletException;
-import javax.servlet.ServletRequest;
-import javax.servlet.ServletResponse;
-import javax.servlet.http.HttpServletRequest;
-import lombok.RequiredArgsConstructor;
-import org.springframework.security.core.Authentication;
-import org.springframework.security.core.context.SecurityContextHolder;
-
-@RequiredArgsConstructor
-public class JwtFilter extends GenericFilter {
-
-  private final TokenProvider tokenProvider;
-
-  @Override
-  public void doFilter(
-      ServletRequest request, ServletResponse response, FilterChain chain)
-      throws IOException, ServletException {
-    HttpServletRequest httpServletRequest = (HttpServletRequest) request;
-    String jwt = tokenProvider.extractToken(httpServletRequest);
-
-    if (jwt != null && tokenProvider.validateToken(jwt)) {
-      Authentication authentication = tokenProvider.getAuthentication(jwt);
-      SecurityContextHolder.getContext().setAuthentication(authentication);
-    }
-    chain.doFilter(request, response);
-  }
-}
