--- server/src/main/java/com/inha/server/security/jwt/JwtSecurityConfig.java
@@ -1,20 +1,0 @@
-package com.inha.server.security.jwt;
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.security.config.annotation.SecurityConfigurerAdapter;
-import org.springframework.security.config.annotation.web.builders.HttpSecurity;
-import org.springframework.security.web.DefaultSecurityFilterChain;
-import org.springframework.security.web.authentication.UsernamePasswordAuthenticationFilter;
-
-@RequiredArgsConstructor
-public class JwtSecurityConfig extends
-    SecurityConfigurerAdapter<DefaultSecurityFilterChain, HttpSecurity> {
-
-  private final TokenProvider tokenProvider;
-
-  @Override
-  public void configure(HttpSecurity httpSecurity) throws Exception {
-    JwtFilter jwtFilter = new JwtFilter(tokenProvider);
-    httpSecurity.addFilterBefore(jwtFilter, UsernamePasswordAuthenticationFilter.class);
-  }
-}
