--- server/src/main/java/com/inha/server/security/jwt/TokenProvider.java
@@ -1,102 +1,0 @@
-package com.inha.server.security.jwt;
-
-import io.jsonwebtoken.Claims;
-import io.jsonwebtoken.ExpiredJwtException;
-import io.jsonwebtoken.Jwts;
-import io.jsonwebtoken.MalformedJwtException;
-import io.jsonwebtoken.SignatureAlgorithm;
-import io.jsonwebtoken.io.Decoders;
-import io.jsonwebtoken.security.Keys;
-import java.security.Key;
-import java.util.Arrays;
-import java.util.Collection;
-import java.util.Date;
-import java.util.stream.Collectors;
-import javax.servlet.http.HttpServletRequest;
-import org.springframework.beans.factory.InitializingBean;
-import org.springframework.beans.factory.annotation.Value;
-import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
-import org.springframework.security.core.Authentication;
-import org.springframework.security.core.GrantedAuthority;
-import org.springframework.security.core.authority.SimpleGrantedAuthority;
-import org.springframework.security.core.userdetails.User;
-import org.springframework.stereotype.Service;
-import org.springframework.util.StringUtils;
-
-@Service
-public class TokenProvider implements InitializingBean {
-
-  private static final String AUTHORIZATION_HEADER = "x-access-token";
-  private static final String AUTHORITY_KEY = "auth";
-  private final String secret;
-  private final Long tokenValidityInSeconds;
-  private Key key;
-
-  public TokenProvider(@Value("${jwt.secret}") String secret,
-      @Value("${jwt.token-validity-in-seconds}") Long tokenValidityInSeconds) {
-    this.secret = secret;
-    this.tokenValidityInSeconds = tokenValidityInSeconds;
-  }
-
-  @Override
-  public void afterPropertiesSet() throws Exception {
-    byte[] keyBytes = Decoders.BASE64.decode(secret);
-    this.key = Keys.hmacShaKeyFor(keyBytes);
-  }
-
-  public String createToken(Authentication authentication) {
-    String authorities = authentication.getAuthorities().stream()
-        .map(GrantedAuthority::getAuthority)
-        .collect(Collectors.joining(","));
-
-    long now = new Date().getTime();
-    Date validity = new Date(now + this.tokenValidityInSeconds);
-
-    return Jwts.builder()
-        .setSubject(authentication.getName())
-        .claim(AUTHORITY_KEY, authorities)
-        .signWith(key, SignatureAlgorithm.HS512)
-        .setExpiration(validity)
-        .compact();
-  }
-
-  public String extractToken(HttpServletRequest request) {
-    String token = request.getHeader(AUTHORIZATION_HEADER);
-    if (StringUtils.hasText(token)) {
-      return token;
-    }
-    return null;
-  }
-
-  public boolean validateToken(String token) {
-    try {
-      Jwts.parserBuilder().setSigningKey(key).build().parseClaimsJws(token);
-      return true;
-    } catch (SecurityException | MalformedJwtException e) {
-      e.printStackTrace();
-      throw new RuntimeException("잘못된 JWT 서명입니다.");
-    } catch (ExpiredJwtException e) {
-      e.printStackTrace();
-      throw new RuntimeException("만료된 토큰입니다.");
-    } catch (IllegalArgumentException e) {
-      e.printStackTrace();
-      throw new RuntimeException("잘못된 JWT 토큰입니다.");
-    }
-  }
-
-  public Authentication getAuthentication(String token) {
-    Claims claims = Jwts.parserBuilder()
-        .setSigningKey(key)
-        .build()
-        .parseClaimsJws(token)
-        .getBody();
-
-    Collection<? extends GrantedAuthority> authorities = Arrays
-        .stream(claims.get(AUTHORITY_KEY).toString().split(","))
-        .map(SimpleGrantedAuthority::new)
-        .collect(Collectors.toList());
-    User principle = new User(claims.getSubject(), "", authorities);
-
-    return new UsernamePasswordAuthenticationToken(principle, token, authorities);
-  }
-}
